+++ examples/ngic/config/dp_config.cfg
@@ -1,7 +1,7 @@
 NeedKNI=true
 CPUList="1-13"
-S1U_PORT_IDX="2"
-SGI_PORT_IDX="3"
+S1U_PORT_IDX="0"
+SGI_PORT_IDX="1"
 S1U_IP="11.8.1.93"
 SGI_IP="13.8.1.93"
 S1UDeviceName="S1Udev"
