+++ rules/adcrule.go
@@ -15,6 +15,17 @@ import (
 	"strconv"
 )
 
+//constants for adc filter
+const (
+	LblNumAdcRules = "NUM_ADC_RULES"
+	LblAdcRule     = "ADC_RULE"
+
+	LblAdcType = "ADC_TYPE"
+	LblIP      = "IP"
+	LblPrefix  = "PREFIX"
+	LblDomain  = "DOMAIN"
+)
+
 //ADCFilter ...
 type ADCFilter struct {
 	ADCFilterIdx int
@@ -36,7 +47,7 @@ func ParseADCRules(filepath string) []ADCFilter {
 
 	noOfFilters, err := cfg.Section(LblGlobal).Key(LblNumAdcRules).Int()
 	flow.CheckFatal(err)
-
+	adcFilters := []ADCFilter{}
 	for i := 1; i <= noOfFilters; i++ {
 		strVal := strconv.Itoa(i)
 		section := cfg.Section(LblAdcRule + "_" + strVal)
